refactor(src/NewProposal): Move NewProposalForm to NewProposalPage
